[API] - Removing process model export & migrating helper function #132
