+++ server/models/cpu.js
@@ -51,41 +51,13 @@ const CpuLogsSchema = new mongoose.Schema({
   timestamp: {
     type: Date,
   },
-  processes: ProcessSchema,
+  arrays: [ProcessSchema],
 })
 
 const CpuLogs = mongoose.model('cpu_logs', CpuLogsSchema)
-
-async function processCpuLogInfo(payload) {
-  //load values
-  const { deviceId, timestamp, processes } = payload
-
-  //compute values
-  const usagePercentage = 0
-  const usageSpeed = 0
-  const numProcesses = processes.length
-  const threadsAlive = 0
-  const threadsSleeping = 0
-  const uptime = 0
-  const newProcesses = new Processes({ processes })
-
-  return new CpuLogs({
-    deviceId,
-    usagePercentage,
-    usageSpeed,
-    numProcesses,
-    threadsAlive,
-    threadsSleeping,
-    uptime,
-    timestamp,
-    processes: newProcesses,
-  })
-}
 
 module.exports = {
   CpuSchema: CpuSchema,
   CpuLogsSchema: CpuLogsSchema,
   CpuLogs: CpuLogs,
-
-  processCpuLogInfo: processCpuLogInfo,
 }
